# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Knowledge Base API",
    "version": "1.0.0",
    "description": "REST API สำหรับจัดการ Knowledge Base - ใช้ x-api-key header สำหรับ authentication"
  },
  "servers": [
    { "url": "http://localhost:3000", "description": "Local" }
  ],
  "security": [{ "ApiKeyAuth": [] }],
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key"
      }
    },
...
  },
  "paths": {
...
    "/api/v1/kb/categories": {
      "get": {
        "summary": "รายการ Categories",
        "operationId": "listCategories",
        "responses": { "200": { "description": "List of categories" } }
      },
      "post": {
        "summary": "สร้าง Category ใหม่",
        "operationId": "createCategory",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name"],
                "properties": {
                  "name": { "type": "string", "maxLength": 100 },
                  "description": { "type": "string", "nullable": true },
                  "sortOrder": { "type": "integer", "default": 0 }
                }
              }
            }
          }
        },
        "responses": { "201": { "description": "Created category" } }
      }
    },
    "/api/v1/kb/categories/{id}": {
      "get": {
        "summary": "ดึง Category ตาม ID",
        "operationId": "getCategory",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "responses": { "200": { "description": "Category" } }
      },
      "put": {
        "summary": "แก้ไข Category",
        "operationId": "updateCategory",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": { "type": "string", "maxLength": 100 },
                  "description": { "type": "string", "nullable": true },
                  "sortOrder": { "type": "integer" }
                }
              }
            }
          }
        },
        "responses": { "200": { "description": "Updated category" } }
      },
      "delete": {
        "summary": "ลบ Category",
        "operationId": "deleteCategory",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "responses": { "200": { "description": "Deleted" } }
      }
    },
...
    "/api/v1/kb/flows/{id}": {
      "get": {
        "summary": "ดึง Flow พร้อม Steps",
        "operationId": "getFlow",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "responses": { "200": { "description": "Flow with steps" } }
      },
      "put": {
        "summary": "แก้ไข Flow",
        "operationId": "updateFlow",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": { "type": "string" },
                  "description": { "type": "string", "nullable": true },
                  "triggerKeywords": { "type": "array", "items": { "type": "string" } },
                  "priority": { "type": "integer" },
                  "isActive": { "type": "boolean" }
                }
              }
            }
          }
        },
        "responses": { "200": { "description": "Updated flow" } }
      },
      "delete": {
        "summary": "ลบ Flow พร้อม Steps",
        "operationId": "deleteFlow",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "responses": { "200": { "description": "Deleted" } }
      }
    },
...
    "/api/v1/kb/escalate": {
      "post": {
        "summary": "สร้าง Escalation ใหม่",
        "operationId": "createEscalation",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["question"],
                "properties": {
                  "question": { "type": "string" },
                  "customerContext": { "type": "string" },
                  "aiAttemptedAnswer": { "type": "string" },
                  "priority": { "type": "string", "enum": ["low", "normal", "high"], "default": "normal" },
                  "sessionId": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": { "200": { "description": "Escalation created" } }
      }
    },
...
    "/api/v1/kb/escalations/{id}/resolve": {
      "patch": {
        "summary": "Resolve Escalation (เพิ่ม KB อัตโนมัติถ้า added_to_kb)",
        "operationId": "resolveEscalation",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["answer", "resolutionType"],
                "properties": {
                  "answer": { "type": "string" },
                  "resolutionType": { "type": "string", "enum": ["answered", "added_to_kb", "redirected", "not_relevant"] },
                  "adminNotes": { "type": "string", "nullable": true },
                  "categoryId": { "type": "string", "format": "uuid", "nullable": true },
                  "keywords": { "type": "array", "items": { "type": "string" } }
                }
              }
            }
          }
        },
        "responses": { "200": { "description": "Resolved" } }
      }
    },
...
  }
}